# api/step-ci/schemas/openapi.yaml
openapi: 3.0.3
info:
  title: Todoist AI Agent API
  description: |
    AI-powered task management API following Google API Design Guide.
    
    ## Authentication
    Currently using environment-based API keys for Todoist and Claude.
    
    ## Resource Naming
    - Resources use kebab-case for URIs
    - Collections are plural (tasks, projects, labels)
    - Custom methods use :verb suffix pattern
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: ISC
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-app.vercel.app/api
    description: Production server (Vercel)
...
paths:
...
  /v1/tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
        description: Task ID
    get:
      tags: [tasks]
      operationId: getTask
      summary: Get a task
      description: Returns a single task by ID
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [tasks]
      operationId: updateTask
      summary: Update a task
      description: Updates task properties
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [tasks]
      operationId: deleteTask
      summary: Delete a task
      description: Deletes a task from storage
      responses:
        '204':
          description: Task deleted
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===========================================================================
  # Tasks - Custom Methods
  # ===========================================================================
...
  /v1/ai/classify:
    post:
      tags: [ai]
      operationId: classifyTasks
      summary: Classify tasks
      description: Use AI to classify tasks into life area categories
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                taskIds:
                  type: array
                  items:
                    type: string
                  description: Specific task IDs to classify (optional)
                force:
                  type: boolean
                  default: false
                  description: Re-classify already classified tasks
                batchSize:
                  type: integer
                  default: 10
                  maximum: 50
      responses:
        '200':
          description: Classification results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassifyTasksResponse'
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, content, projectId, priority, createdAt, updatedAt, isCompleted]
      properties:
        id:
          type: string
        content:
          type: string
        description:
          type: string
        projectId:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 4
        labels:
          type: array
          items:
            type: string
        due:
          $ref: '#/components/schemas/TaskDue'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isCompleted:
          type: boolean
        completedAt:
          type: string
          format: date-time
        metadata:
          $ref: '#/components/schemas/TaskMetadata'
    TaskDue:
      type: object
      properties:
        date:
          type: string
          format: date
        datetime:
          type: string
          format: date-time
        string:
          type: string
        isRecurring:
          type: boolean
    TaskMetadata:
      type: object
      properties:
        category:
          type: string
        timeEstimate:
          type: string
        timeEstimateMinutes:
          type: integer
        size:
          type: string
          enum: [XS, S, M, L, XL]
        aiConfidence:
          type: number
          minimum: 0
          maximum: 1
...
    UpdateTaskRequest:
      type: object
      properties:
        content:
          type: string
        description:
          type: string
        projectId:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 4
        category:
          type: string
        labels:
          type: array
          items:
            type: string
...
    ClassifyTasksResponse:
      type: object
      required: [results, tasksProcessed, tasksClassified, duration]
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              taskId:
                type: string
              taskContent:
                type: string
              category:
                type: string
              confidence:
                type: number
              reasoning:
                type: string
        tasksProcessed:
          type: integer
        tasksClassified:
          type: integer
        duration:
          type: integer
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, status]
          properties:
            code:
              type: integer
            message:
              type: string
            status:
              type: string
            details:
              type: array
              items:
                type: object